# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MailDay Matching API
servers:
  - url: /api
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: Get match history
      parameters:
        - name: age_band
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Match history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithChildren"
    post:
      operationId: approveMatch
      tags: [matches]
      summary: Approve a match suggestion (creates match record)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchApproval"
      responses:
        "201":
          description: Match approved and created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a single match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithChildren"
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update a match (e.g., close it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
...
components:
  schemas:
...
    Parent:
      type: object
      required: [id, first_name, last_name, email]
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        mailing_address:
          type: ["string", "null"]
        address_type:
          type: ["string", "null"]
        shopify_customer_id:
          type: ["string", "null"]
        recharge_subscription_id:
          type: ["string", "null"]
        membership_tier:
          type: ["string", "null"]
        billing_type:
          type: ["string", "null"]
        subscription_status:
          type: ["string", "null"]
        join_date:
          type: ["string", "null"]
        referral_source:
          type: ["string", "null"]
        community_status:
          type: ["string", "null"]
        give_a_key_recipient:
          type: ["boolean", "null"]
        last_email_open_date:
          type: ["string", "null"]
        at_risk:
          type: ["boolean", "null"]
        internal_notes:
          type: ["string", "null"]
        created_at:
          type: string
...
    Child:
      type: object
      required: [id, parent_id, child_first_name, age, tier, match_status]
      properties:
        id:
          type: string
        parent_id:
          type: string
        child_first_name:
          type: string
        age:
          type: integer
        tier:
          type: string
          description: Core / Minis / Homeschool Core / Homeschool Minis
        interests:
          type: array
          items:
            type: string
        homeschool_edition:
          type: ["boolean", "null"]
        homeschool_tier:
          type: ["string", "null"]
        homeschool_approach:
          type: ["string", "null"]
        match_status:
          type: string
          description: Unmatched / Matched / Rematch Requested / Rematched / Paused / Cancelled
        rematch_count:
          type: ["integer", "null"]
        match_guarantee_start_date:
          type: ["string", "null"]
        billing_paused:
          type: ["boolean", "null"]
        safety_flag:
          type: ["boolean", "null"]
        internal_notes:
          type: ["string", "null"]
        created_date:
          type: ["string", "null"]
        created_at:
          type: string
...
    ChildWithParent:
      allOf:
        - $ref: "#/components/schemas/Child"
        - type: object
          required: [parent]
          properties:
            parent:
              $ref: "#/components/schemas/Parent"
            days_waiting:
              type: ["integer", "null"]
            guarantee_status:
              type: ["string", "null"]
              description: ok / warning / urgent
...
    MatchApproval:
      type: object
      required: [child_a_id, child_b_id, shared_interests]
      properties:
        child_a_id:
          type: string
        child_b_id:
          type: string
        shared_interests:
          type: array
          items:
            type: string
    Match:
      type: object
      required: [id, child_a_id, child_b_id, match_status, created_at]
      properties:
        id:
          type: string
        child_a_id:
          type: string
        child_b_id:
          type: string
        match_date:
          type: ["string", "null"]
        match_status:
          type: string
          description: Active / Closed
        shared_interests:
          type: array
          items:
            type: string
        match_notification_sent_a:
          type: ["boolean", "null"]
        match_notification_sent_b:
          type: ["boolean", "null"]
        close_reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
    MatchUpdate:
      type: object
      properties:
        match_status:
          type: string
        close_reason:
          type: string
        notes:
          type: string
        match_notification_sent_a:
          type: boolean
        match_notification_sent_b:
          type: boolean
    MatchWithChildren:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            child_a:
              $ref: "#/components/schemas/ChildWithParent"
            child_b:
              $ref: "#/components/schemas/ChildWithParent"
...